Avoid using shadowy techniques to get around a package private interface
