--- src/main/java/net/minecraft/util/shape/Lithium_DoubleListPair.java
@@ -1,6 +1,0 @@
-package net.minecraft.util.shape;
-
-// Interface to expose the super-interface.
-public interface Lithium_DoubleListPair extends DoubleListPair {
-
-}
